Layout fixes and media player initialization fixes
